dataStore 관련 코드 refactor, packaging
flow -> suspend
동일한 로직을 가지는 이해하기 쉬운 함수로 변경
