+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/datasource/GuestLoginTokenDataSource.kt
@@ -1,8 +1,6 @@
 package com.nexters.bandalart.android.core.data.datasource
 
-import kotlinx.coroutines.flow.Flow
-
 interface GuestLoginTokenDataSource {
   suspend fun setGuestLoginToken(guestLoginToken: String)
-  fun getGuestLoginToken(): Flow<String>
+  suspend fun getGuestLoginToken(): String
 }
