+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/di/DataStoreProviderModule.kt
@@ -2,9 +2,8 @@ package com.nexters.bandalart.android.core.data.di
 
 import android.content.Context
 import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStoreFile
+import androidx.datastore.preferences.preferencesDataStore
 import com.nexters.bandalart.android.core.data.local.DataStoreProvider
 import dagger.Module
 import dagger.Provides
@@ -13,7 +12,8 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
-private const val DATASTORE_PREFERENCES = "datastore_preferences"
+private const val TOKEN_DATASTORE = "token_datastore"
+private val Context.tokenDataStore: DataStore<Preferences> by preferencesDataStore(name = TOKEN_DATASTORE)
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -21,8 +21,7 @@ object DataStoreProviderModule {
 
   @Provides
   @Singleton
-  fun providePreferencesDataStore(@ApplicationContext context: Context): DataStore<Preferences> =
-    PreferenceDataStoreFactory.create(produceFile = { context.preferencesDataStoreFile(DATASTORE_PREFERENCES) })
+  fun providePreferencesDataStore(@ApplicationContext context: Context) = context.tokenDataStore
 
   @Provides
   @Singleton
