+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/local/DataStoreProvider.kt
@@ -7,12 +7,12 @@ import androidx.datastore.preferences.core.emptyPreferences
 import androidx.datastore.preferences.core.stringPreferencesKey
 import java.io.IOException
 import javax.inject.Inject
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
 
-class DataStoreProvider @Inject constructor(private val dataStore: DataStore<Preferences>) {
+class DataStoreProvider @Inject constructor(
+  private val dataStore: DataStore<Preferences>
+) {
 
   companion object {
     private const val GUEST_LOGIN_TOKEN = "guest_login_token"
@@ -28,17 +28,11 @@ class DataStoreProvider @Inject constructor(private val dataStore: DataStore<Pre
     }
   }
 
-  fun getGuestLoginToken(): Flow<String> {
-    return dataStore.data.catch { exception ->
-      if (exception is IOException) {
-        emit(emptyPreferences())
-      } else {
-        throw exception
-      }
-    }.map { preference ->
-      preference[prefKeyGuestLoginToken] ?: ""
-    }
-  }
+  suspend fun getGuestLoginToken() = dataStore.data
+    .catch { exception ->
+      if (exception is IOException) emit(emptyPreferences())
+      else throw exception
+    }.first()[prefKeyGuestLoginToken] ?: ""
 
   suspend fun setRecentBandalartKey(recentBandalartKey: String) {
     dataStore.edit { preferences ->
