+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/local/datasource/GuestLoginTokenDataSourceImpl.kt
@@ -3,7 +3,6 @@ package com.nexters.bandalart.android.core.data.local.datasource
 import com.nexters.bandalart.android.core.data.datasource.GuestLoginTokenDataSource
 import com.nexters.bandalart.android.core.data.local.DataStoreProvider
 import javax.inject.Inject
-import kotlinx.coroutines.flow.Flow
 
 class GuestLoginTokenDataSourceImpl @Inject constructor(private val datastoreProvider: DataStoreProvider) :
   GuestLoginTokenDataSource {
@@ -12,7 +11,7 @@ class GuestLoginTokenDataSourceImpl @Inject constructor(private val datastorePro
     datastoreProvider.setGuestLoginToken(guestLoginToken)
   }
 
-  override fun getGuestLoginToken(): Flow<String> {
+  override suspend fun getGuestLoginToken(): String {
     return datastoreProvider.getGuestLoginToken()
   }
 }
