+++ core/data/src/main/kotlin/com/nexters/bandalart/android/core/data/repository/GuestLoginTokenRepositoryImpl.kt
@@ -3,7 +3,6 @@ package com.nexters.bandalart.android.core.data.repository
 import com.nexters.bandalart.android.core.data.datasource.GuestLoginTokenDataSource
 import com.nexters.bandalart.android.core.domain.repository.GuestLoginTokenRepository
 import javax.inject.Inject
-import kotlinx.coroutines.flow.Flow
 
 class GuestLoginTokenRepositoryImpl @Inject constructor(private val dataSource: GuestLoginTokenDataSource) :
   GuestLoginTokenRepository {
@@ -11,7 +10,7 @@ class GuestLoginTokenRepositoryImpl @Inject constructor(private val dataSource: 
     dataSource.setGuestLoginToken(guestLoginToken)
   }
 
-  override fun getGuestLoginToken(): Flow<String> {
+  override suspend fun getGuestLoginToken(): String {
     return dataSource.getGuestLoginToken()
   }
 }
