+++ core/domain/src/main/kotlin/com/nexters/bandalart/android/core/domain/repository/GuestLoginTokenRepository.kt
@@ -1,7 +1,5 @@
 package com.nexters.bandalart.android.core.domain.repository
 
-import kotlinx.coroutines.flow.Flow
-
 interface GuestLoginTokenRepository {
   /**
    * 게스트 로그인 토큰 저장
@@ -13,5 +11,5 @@ interface GuestLoginTokenRepository {
   /**
    * 게스트 로그인 토큰 조회
    */
-  fun getGuestLoginToken(): Flow<String>
+  suspend fun getGuestLoginToken(): String
 }
