+++ core/domain/src/main/kotlin/com/nexters/bandalart/android/core/domain/usecase/bandalart/GetRecentBandalartKeyUseCase.kt
@@ -1,4 +1,4 @@
-package com.nexters.bandalart.android.core.domain.usecase
+package com.nexters.bandalart.android.core.domain.usecase.bandalart
 
 import com.nexters.bandalart.android.core.domain.repository.BandalartRepository
 import javax.inject.Inject
